Add Engine Render Hooks
