+++ Modules/@babylonjs/react-native/EngineView.tsx
@@ -1,8 +1,9 @@
-import React, { Component, FunctionComponent, SyntheticEvent, useCallback, useEffect, useState, useRef } from 'react';
-import { requireNativeComponent, ViewProps, AppState, AppStateStatus, View, Text, findNodeHandle, UIManager } from 'react-native';
+import React, { Component, FunctionComponent, SyntheticEvent, useCallback, useEffect, useState, useRef, useMemo } from 'react';
+import { requireNativeComponent, ViewProps, View, Text, findNodeHandle, UIManager } from 'react-native';
 import { Camera, SceneInstrumentation } from '@babylonjs/core';
 import { ensureInitialized } from './BabylonModule';
 import { ReactNativeEngine } from './ReactNativeEngine';
+import { useRenderLoop } from './EngineHook';
 
 declare const global: any;
 
@@ -34,7 +35,6 @@ interface SceneStats {
 
 export const EngineView: FunctionComponent<EngineViewProps> = (props: EngineViewProps) => {
     const [initialized, setInitialized] = useState<boolean>();
-    const [appState, setAppState] = useState(AppState.currentState);
     //const [fps, setFps] = useState<number>();
     const [sceneStats, setSceneStats] = useState<SceneStats>();
     const engineViewRef = useRef<Component<NativeEngineViewProps>>(null);
@@ -47,39 +47,17 @@ export const EngineView: FunctionComponent<EngineViewProps> = (props: EngineView
         })();
     }, []);
 
-    useEffect(() => {
-        const onAppStateChanged = (appState: AppStateStatus) => {
-            setAppState(appState);
-        };
+    const engine = useMemo(() => {
+        return props.camera?.getScene().getEngine() as ReactNativeEngine;
+    }, [props.camera]);
 
-        AppState.addEventListener("change", onAppStateChanged);
-
-        return () => {
-            AppState.removeEventListener("change", onAppStateChanged);
+    const renderLoop = useCallback(() => {
+        for (let scene of engine.scenes) {
+            scene.render();
         }
-    }, []);
+    }, [engine]);
 
-    useEffect(() => {
-        if (props.camera && appState === "active") {
-            const engine = props.camera.getScene().getEngine() as ReactNativeEngine;
-
-            if (!engine.isDisposed) {
-                engine.runRenderLoop(() => {
-                    for (let scene of engine.scenes) {
-                        scene.render();
-                    }
-                });
-
-                return () => {
-                    if (!engine.isDisposed) {
-                        engine.stopRenderLoop();
-                    }
-                };
-            }
-        }
-
-        return undefined;
-    }, [props.camera, appState]);
+    useRenderLoop(engine, renderLoop);
 
     useEffect(() => {
         if (props.camera && (props.displayFrameRate ?? __DEV__)) {
